fixed-sidebar-routing-issue
