+++ src/app/(root)/display/display.tsx
