+++ src/app/(root)/notifications/page.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const page = () => {
+  return (
+    <div>Notifications</div>
+  )
+}
+
+export default page
