+++ src/app/(root)/page.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const page = () => {
+    return (
+        <div>Dashboard</div>
+    )
+}
+
+export default page
