+++ src/app/(root)/profile/layout.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+
+
+export default function Layout({
+    children,
+}: Readonly<{
+    children: React.ReactNode;
+}>) {
+    return (
+
+        <div className="">
+            {children}
+        </div>
+    );
+}
