+++ src/app/(root)/profile/page.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const page = () => {
+  return (
+    <div>Profile</div>
+  )
+}
+
+export default page
